Removed NLTK
Using the NLTK to find word synonyms was a good idea on paper but in practice it produced very few results and mostly only caused a slowdown of nearly 50%. It may be a good consideration in the future, but other alternatives may prove to be more efficient and produce better results.

pySpellCheck is still being used. Despite a few obvious errors it's better than nothing for now and does not negatively impact performance considerably.