Refactored (Almost) Everything
Reorganized a lot of the functions / files in an attempt to make everything easier to understand.

I also wanted to make generateGraph.py less monolithic. Since the most complex part of it was finding the edges, I moved that portion of the code to another file. I also added a separate edge finder for programming files since the original (now used only for DB files) was giving incorrect output.